requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
aiogram>=3.6
sqlalchemy>=2.0
asyncpg>=0.29
httpx>=0.27
pydantic>=2.4,<3.0
pydantic-settings>=2.0
python-dotenv>=1.0
Pillow>=10.0
loguru>=0.7
beautifulsoup4==4.12.3
selectolax